SQUASH_ME: QML stuff ported
