+++ package/metadata.json
@@ -1,4 +1,5 @@
 {
+    "KPackageStructure": "Plasma/Applet",
     "KPlugin": {
         "Authors": [
             {
@@ -10,16 +11,11 @@
         "Description": "Small example to showcase a Plasmoid with a Rust backend",
         "EnabledByDefault": true,
         "Icon": "applications-system",
-        "Id": "org.kde.plasma.rust.example",
+        "Id": "org.kde.plasma.rust_backend_example",
         "License": "MPL 2.0",
         "Name": "Rust backend example",
-        "ServiceTypes": [
-            "Plasma/Applet"
-        ],
         "Version": "1.0",
-        "Website": "TODO"
+        "Website": "https://github.com/brummer-simon/rust_plasmoid_example"
     },
-    "KPackageStructure": "Plasma/Applet",
-    "X-Plasma-API": "declarativeappletscript",
-    "X-Plasma-MainScript": "ui/main.qml"
+    "X-Plasma-API-Minimum-Version": "6.0"
 }
