+++ plugin/Cargo.toml
@@ -12,4 +12,4 @@ name = "rust_backend"
 crate-type = ["cdylib"]
 
 [dependencies]
-qmetaobject = { version = "0.2.9"}
+qmetaobject = { version = "0.2.10"}
