+++ scripts/uninstall.sh
@@ -1,0 +1,4 @@
+# TODO: Implement me
+# TODO: Add make file?
+# /usr/share/plasma/plasmoids/org.kde.plasma.rust_backend.example
+# /usr/share/metainfo/org.kde.plasma.rust_backend.example.appdata.xml
